timers: DOM types
